README
Not all data is included here. Data coming from the CDISC Pilot Study can be found at the CDISC website. 

Data from the CDISC Pilot Study that are not included are: ADSL, ADLBC, ADAE, ADTTE, ADVS.

Note that ADAE found in this repository is not the same ADAE used for the CDISC Pilot Study. ADAE found in this repository is used for Chapter 8 only.  For Chapter 2 and Chapter 7, use ADAE from the CDISC Pilot Study.